libutil/read_all: rename from readall.[ch]
Problem: readall.[ch] contains function named read_all().

Rename readall.[ch] to read_all.[ch].

Update users.
